# public/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "etiennetran@hotmail.fr"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get list of articles",
                "description": "Get all articles.",
                "operationId": "getAllArticles",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number per page"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Current page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Article"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "scalar": "api_key"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Create an article",
                "description": "Create an article.",
                "operationId": "createArticle",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Article"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Article"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "scalar": "api_key"
                    }
                ]
            }
        },
        "/articles/{articleId}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get the article.",
                "description": "Get the article.",
                "operationId": "getArticleById",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "article id",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Article"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "scalar": "api_key"
                    }
                ]
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Update the article.",
                "description": "Update the article.",
                "operationId": "updateArticle",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "article id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Article"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Article"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "scalar": "api_key"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Delete the article.",
                "description": "Delete the article.",
                "operationId": "deleteArticleById",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "article id",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "scalar": "api_key"
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "description": "Get all users.",
                "operationId": "getAllUsers",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "scalar": "api_key"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a user",
                "description": "Create a user.",
                "operationId": "createUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    }
                }
            }
        },
        "/users/{userId}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get the user",
                "description": "Get the user.",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "user id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "scalar": "api_key"
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a user",
                "description": "Update a user.",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "user id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "scalar": "api_key"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete the user.",
                "description": "Delete the user.",
                "operationId": "deleteUserById",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "user id",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "scalar": "api_key"
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Links": {
                "properties": {
                    "rel": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "GET"
                        ]
                    },
                    "href": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Article": {
                "description": "App\\Models\\Article",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "rate": {
                        "type": "string"
                    },
                    "format": {
                        "type": "string",
                        "enum": [
                            "dvd",
                            "blu-ray"
                        ]
                    },
                    "trailerUrl": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "releaseDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "_links": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Links"
                        },
                        "readOnly": true
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "_links": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Links"
                        },
                        "readOnly": true
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "NotFound": {
                "description": "Resource not found.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "404": {
                                    "type": "string",
                                    "example": "The article does not exist."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "BadRequest": {
                "description": "Bad request.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "400": {
                                    "type": "string",
                                    "example": "Bad request."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "Unauthorized.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "401": {
                                    "type": "string",
                                    "example": "Unauthenticated."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "api_token",
                "in": "query"
            }
        }
    }
}